embeddings_service/requirements.txt
grpcio-tools
sentence-transformers==2.2.2